Move tag mapping logic into a util function
