+++ src/utils/card.ts
@@ -1,6 +1,6 @@
 import { Card, ModifiedCard, TAG } from '../types/card.ts'
 
-export function sortEffectCardsFirst(a: Card, b: Card) {
+export const sortEffectCardsFirst = (a: Card, b: Card) => {
 	if (!!a.effect === !!b.effect) {
 		return 0
 	}
@@ -10,11 +10,11 @@ export function sortEffectCardsFirst(a: Card, b: Card) {
 	return 1
 }
 
-export function findCard<T extends Card>(hand: T[], id: number) {
+export const findCard = <T extends Card>(hand: T[], id: number) => {
 	return hand.find(card => card.id === id) as T
 }
 
-export function removeTag(card: ModifiedCard, tag: TAG, count: number = 1) {
+export const removeTag = (card: ModifiedCard, tag: TAG, count: number = 1) => {
 	let deletedCount = 0
 	const finalTags = card.modifiedTags.reduce<TAG[]>((acc, t) => {
 		if (t === tag && deletedCount < count) {
@@ -28,4 +28,20 @@ export function removeTag(card: ModifiedCard, tag: TAG, count: number = 1) {
 		throw new Error('Not enough tags to remove')
 	}
 	card.modifiedTags = finalTags
+}
+
+export const mapCardTags = (card: ModifiedCard) => {
+	const addedTags = card.modifiedTags.slice().sort((a, b) => a - b)
+	const originalTagList = card.isTransformed && card.transformedTags ? card.transformedTags : card.tags
+	const tags = originalTagList.map(t => ({ tag: t, isDeleted: false }))
+	for (const tag of tags) {
+		const resultIndex = addedTags.indexOf(tag.tag)
+		if (resultIndex > -1) {
+			addedTags.splice(resultIndex, 1)
+		} else {
+			tag.isDeleted = true
+		}
+	}
+
+	return [tags, addedTags] as const
 }
